Se actualiza a ultima version con Pauta Call Voz V2 (v4 en interno) y modales nuevos en vistas de asignaciones para ver datos de centro y comentario estado.

+++ database/migrations/2021_01_12_204535_create_atributos_table.php
@@ -27,7 +27,7 @@ class CreateAtributosTable extends Migration
             $table->boolean('no_aplica')->nullable()->default(false);
             $table->string('name_categoria');
             $table->integer('ponderador')->nullable()->default(1);
-            $table->integer('ponderador_ici')->nullable()->default(1);
+            $table->float('ponderador_ici')->nullable()->default(1);
             $table->string('tipo_respuesta')->nullable();
             $table->unsignedBigInteger('pauta_id')->nullable();
             $table->foreign('pauta_id')->references('id')->on('pautas')->onDelete('set null');
